<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

<!-- Explain the purpose of this PR and the goals it aims to achieve. -->
This PR is to introduce block scan warp scan in fused MoE path **moe_align_block_size_kernel** .

Here is the benchmark result.
Note: num_experts >= 128 appliable to this PR. num_experts < 128 appliable to **moe_align_block_size_small_batch_expert_kernel**  kernel.

Main:
```
$python ./sgl-kernel/benchmark/bench_moe_align_block_size.py
INFO 07-09 13:15:28 [__init__.py:244] Automatically detected platform cuda.
✅ VLLM implementation works with 64 experts!
✅ SGL and Triton implementations match
✅ SGL and VLLM implementations match

📊 Running performance benchmark for 64 experts...
moe-align-block-size-performance:
     num_tokens  num_experts  topk        SGL  SGL Fusion       Triton
0           1.0          8.0   1.0  16.672000   14.912000    44.032000
1           1.0          8.0   2.0  16.896000   15.072000    44.256002
2           1.0          8.0   4.0  16.672000   15.008000    43.264002
3           1.0          8.0   8.0  16.432000   15.104000    44.767998
4           1.0         32.0   1.0  19.200001   17.424000    44.512000
5           1.0         32.0   2.0  19.231999   17.503999    44.464000
6           1.0         32.0   4.0  19.200001   17.535999    43.728001
7           1.0         32.0   8.0  19.231999   17.568000    44.767998
8           1.0         64.0   1.0  22.655999   20.864001    44.160001
9           1.0         64.0   2.0  22.608001   20.927999    44.799998
10          1.0         64.0   4.0  22.624001   20.927999    44.863999
11          1.0         64.0   8.0  22.496000   20.959999    45.343999
12          1.0        128.0   1.0  19.296000   17.696001    47.295999
13          1.0        128.0   2.0  19.264000   17.472001    44.927999
14          1.0        128.0   4.0  19.328000   17.728001    43.968000
15          1.0        128.0   8.0  19.296000   17.503999    44.560000
16          1.0        256.0   1.0  19.231999   17.632000    45.311999
17          1.0        256.0   2.0  19.200001   17.728001    45.056000
18          1.0        256.0   4.0  19.360000   17.440001    46.271998
19          1.0        256.0   8.0  19.328000   17.759999    46.016000
20          8.0          8.0   1.0  16.543999   15.104000    44.096000
21          8.0          8.0   2.0  16.576000   14.976000    44.383999
22          8.0          8.0   4.0  16.319999   15.008000    45.407999
23          8.0          8.0   8.0  17.055999   15.488000    44.911999
24          8.0         32.0   1.0  19.216000   17.568000    45.152001
25          8.0         32.0   2.0  19.231999   17.728001    45.919999
26          8.0         32.0   4.0  19.072000   17.824000    45.887999
27          8.0         32.0   8.0  19.776000   18.528000    45.664001
28          8.0         64.0   1.0  22.608001   20.959999    46.112001
29          8.0         64.0   2.0  22.528000   20.992000    43.584000
30          8.0         64.0   4.0  22.624001   21.183999    44.704001
31          8.0         64.0   8.0  22.496000   21.376001    45.664001
32          8.0        128.0   1.0  19.328000   17.535999    46.464000
33          8.0        128.0   2.0  19.360000   17.696001    44.831999
34          8.0        128.0   4.0  19.328000   17.759999    46.592001
35          8.0        128.0   8.0  19.360000   17.792000    45.279998
36          8.0        256.0   1.0  19.296000   17.535999    45.823999
37          8.0        256.0   2.0  19.360000   17.568000    46.416000
38          8.0        256.0   4.0  19.328000   17.696001    46.271998
39          8.0        256.0   8.0  19.360000   17.983999    46.048000
40         16.0          8.0   1.0  16.559999   15.008000    44.544000
41         16.0          8.0   2.0  16.287999   15.008000    44.895999
42         16.0          8.0   4.0  17.055999   15.456000    44.576000
43         16.0          8.0   8.0  18.144000   16.543999    45.600001
44         16.0         32.0   1.0  19.231999   17.759999    43.807998
45         16.0         32.0   2.0  19.296000   17.856000    45.248002
46         16.0         32.0   4.0  19.552000   18.480000    45.056000
47         16.0         32.0   8.0  20.832000   19.584000    44.512000
48         16.0         64.0   1.0  22.624001   20.992000    44.927999
49         16.0         64.0   2.0  22.624001   21.120001    47.200002
50         16.0         64.0   4.0  22.592001   21.376001    47.648001
51         16.0         64.0   8.0  22.944000   22.175999    45.728002
52         16.0        128.0   1.0  19.360000   17.535999    45.311999
53         16.0        128.0   2.0  19.392001   17.568000    44.767998
54         16.0        128.0   4.0  19.360000   17.920000    45.472000
55         16.0        128.0   8.0  19.360000   18.015999    45.375999
56         16.0        256.0   1.0  19.328000   17.568000    46.208002
57         16.0        256.0   2.0  19.296000   17.728001    46.944000
58         16.0        256.0   4.0  19.328000   17.983999    46.815999
59         16.0        256.0   8.0  19.424001   18.368000    46.432000
60         32.0          8.0   1.0  16.543999   15.040000    43.744002
61         32.0          8.0   2.0  16.799999   15.488000    44.064000
62         32.0          8.0   4.0  18.112000   16.576000    43.200001
63         32.0          8.0   8.0  20.256000   18.816000    42.495999
64         32.0         32.0   1.0  19.296000   17.792000    41.951999
65         32.0         32.0   2.0  19.568000   18.495999    41.471999
66         32.0         32.0   4.0  20.576000   19.616000    42.495999
67         32.0         32.0   8.0  22.879999   21.728000    44.128001
68         32.0         64.0   1.0  22.560000   21.183999    43.584000
69         32.0         64.0   2.0  22.560000   21.360001    44.831999
70         32.0         64.0   4.0  22.944000   22.175999    43.935999
71         32.0         64.0   8.0  23.840001   23.200000    45.311999
72         32.0        128.0   1.0  19.328000   17.503999    43.488000
73         32.0        128.0   2.0  19.360000   17.983999    44.256002
74         32.0        128.0   4.0  19.392001   17.983999    43.359999
75         32.0        128.0   8.0  19.552000   18.528000    45.136001
76         32.0        256.0   1.0  19.360000   17.744000    45.984000
77         32.0        256.0   2.0  19.312000   17.759999    45.632001
78         32.0        256.0   4.0  19.360000   18.208001    46.144001
79         32.0        256.0   8.0  19.552000   18.880000    46.528000
80         64.0          8.0   1.0  17.055999   15.488000    44.544000
81         64.0          8.0   2.0  17.856000   16.543999    45.056000
82         64.0          8.0   4.0  20.352000   18.848000    44.064000
83         64.0          8.0   8.0  24.480000   22.784000    45.120001
84         64.0         32.0   1.0  19.743999   18.495999    46.560001
85         64.0         32.0   2.0  20.576000   19.616000    43.520000
86         64.0         32.0   4.0  22.879999   21.695999    43.488000
87         64.0         32.0   8.0  26.208000   25.280001    45.024000
88         64.0         64.0   1.0  22.544000   21.376001    45.040000
89         64.0         64.0   2.0  23.135999   22.175999    44.544000
90         64.0         64.0   4.0  23.840001   23.167999    45.248002
91         64.0         64.0   8.0  26.272001   26.272001    45.456000
92         64.0        128.0   1.0  19.360000   17.759999    44.608001
93         64.0        128.0   2.0  19.392001   17.983999    45.120001
94         64.0        128.0   4.0  19.584000   18.656000    46.879999
95         64.0        128.0   8.0  19.424001   18.592000    46.016000
96         64.0        256.0   1.0  19.264000   17.952001    45.855999
97         64.0        256.0   2.0  19.360000   18.400000    46.303999
98         64.0        256.0   4.0  19.487999   18.688001    46.656001
99         64.0        256.0   8.0  19.664001   19.168001    47.359999
100       128.0          8.0   1.0  17.856000   16.543999    46.528000
101       128.0          8.0   2.0  20.191999   18.848000    45.279998
102       128.0          8.0   4.0  24.512000   22.784000    44.032000
103       128.0          8.0   8.0  19.231999   17.472001    46.080001
104       128.0         32.0   1.0  20.832000   19.616000    45.311999
105       128.0         32.0   2.0  22.848001   21.728000    44.480000
106       128.0         32.0   4.0  26.400000   25.312001    45.120001
107       128.0         32.0   8.0  19.104000   17.360000    43.839999
108       128.0         64.0   1.0  23.072001   22.175999    44.831999
109       128.0         64.0   2.0  23.936000   23.167999    44.799998
110       128.0         64.0   4.0  26.272001   26.272001    44.895999
111       128.0         64.0   8.0  19.632000   18.080000    45.375999
112       128.0        128.0   1.0  19.392001   18.015999    45.024000
113       128.0        128.0   2.0  19.552000   18.624000    45.216002
114       128.0        128.0   4.0  19.424001   18.495999    44.415999
115       128.0        128.0   8.0  19.360000   18.495999    44.496000
116       128.0        256.0   1.0  19.424001   18.432001    45.919999
117       128.0        256.0   2.0  19.504000   18.880000    46.688002
118       128.0        256.0   4.0  19.776000   19.168001    47.488000
119       128.0        256.0   8.0  19.808000   19.776000    48.016001
120       256.0          8.0   1.0  20.223999   18.880000    45.248002
121       256.0          8.0   2.0  24.480000   22.784000    45.088001
122       256.0          8.0   4.0  19.231999   17.503999    46.592001
123       256.0          8.0   8.0  19.296000   17.728001    69.696002
124       256.0         32.0   1.0  22.848001   21.695999    43.648001
125       256.0         32.0   2.0  26.400000   25.312001    44.704001
126       256.0         32.0   4.0  19.040000   17.360000    44.000000
127       256.0         32.0   8.0  19.680001   18.080000    44.480000
128       256.0         64.0   1.0  23.871999   23.167999    44.287998
129       256.0         64.0   2.0  26.432000   26.272001    43.935999
130       256.0         64.0   4.0  19.616000   18.015999    44.000000
131       256.0         64.0   8.0  20.096000   18.751999    42.624000
132       256.0        128.0   1.0  19.616000   18.592000    45.375999
133       256.0        128.0   2.0  19.424001   18.528000    43.552000
134       256.0        128.0   4.0  19.360000   18.528000    45.375999
135       256.0        128.0   8.0  20.096000   19.264000    44.592001
136       256.0        256.0   1.0  19.455999   18.751999    46.432000
137       256.0        256.0   2.0  19.680001   19.584000    47.520000
138       256.0        256.0   4.0  19.840000   19.792000    47.839999
139       256.0        256.0   8.0  20.160001   20.000000    50.080001
140       512.0          8.0   1.0  24.351999   22.784000    42.656001
141       512.0          8.0   2.0  19.200001   17.472001    45.728002
142       512.0          8.0   4.0  19.296000   17.759999    69.728002
143       512.0          8.0   8.0  21.024000   19.296000   114.335999
144       512.0         32.0   1.0  26.400000   25.280001    45.311999
145       512.0         32.0   2.0  19.072000   17.376000    43.712001
146       512.0         32.0   4.0  19.680001   18.048000    44.767998
147       512.0         32.0   8.0  21.056000   19.455999    49.311999
148       512.0         64.0   1.0  26.432000   26.272001    45.279998
149       512.0         64.0   2.0  19.648001   18.255999    45.375999
150       512.0         64.0   4.0  20.064000   18.784000    43.455999
151       512.0         64.0   8.0  21.695999   20.191999    43.423999
152       512.0        128.0   1.0  19.392001   18.592000    45.120001
153       512.0        128.0   2.0  19.360000   18.560000    43.104000
154       512.0        128.0   4.0  20.160001   19.264000    43.616001
155       512.0        128.0   8.0  21.183999   20.352000    45.504000
156       512.0        256.0   1.0  19.696000   19.200001    47.168002
157       512.0        256.0   2.0  19.840000   19.727999    47.488000
158       512.0        256.0   4.0  20.128001   20.160001    50.303999
159       512.0        256.0   8.0  21.792000   21.632001    54.752000
160      1024.0          8.0   1.0  19.231999   17.535999    46.239998
161      1024.0          8.0   2.0  19.264000   17.664000    69.824003
162      1024.0          8.0   4.0  20.992000   19.392001   114.944004
163      1024.0          8.0   8.0  23.072001   21.568000   203.743994
164      1024.0         32.0   1.0  19.072000   17.376000    44.112001
165      1024.0         32.0   2.0  19.680001   18.015999    42.176001
166      1024.0         32.0   4.0  21.056000   19.455999    48.864000
167      1024.0         32.0   8.0  23.936000   22.688000    73.824003
168      1024.0         64.0   1.0  19.648001   18.239999    44.096000
169      1024.0         64.0   2.0  20.096000   18.751999    43.776002
170      1024.0         64.0   4.0  21.695999   20.240000    45.056000
171      1024.0         64.0   8.0  24.160000   22.976000    53.215999
172      1024.0        128.0   1.0  19.360000   18.464001    45.216002
173      1024.0        128.0   2.0  20.176000   19.296000    43.807998
174      1024.0        128.0   4.0  21.215999   20.352000    44.447999
175      1024.0        128.0   8.0  23.391999   22.368001    47.839999
176      1024.0        256.0   1.0  19.840000   19.776000    48.416000
177      1024.0        256.0   2.0  20.128001   20.048000    50.080001
178      1024.0        256.0   4.0  21.824000   21.695999    54.719999
179      1024.0        256.0   8.0  24.095999   24.160000    60.640000
180      2048.0          8.0   1.0  19.296000   17.744000    69.888003
181      2048.0          8.0   2.0  20.992000   19.328000   114.432000
182      2048.0          8.0   4.0  23.104001   21.568000   203.968003
183      2048.0          8.0   8.0  27.775999   26.912000   378.239989
184      2048.0         32.0   1.0  19.680001   17.983999    44.128001
185      2048.0         32.0   2.0  21.056000   19.424001    48.448000
186      2048.0         32.0   4.0  23.840001   22.496000    73.600002
187      2048.0         32.0   8.0  28.672000   28.160000   118.207999
188      2048.0         64.0   1.0  20.032000   18.751999    43.968000
189      2048.0         64.0   2.0  21.728000   20.352000    43.903999
190      2048.0         64.0   4.0  24.127999   23.072001    53.183999
191      2048.0         64.0   8.0  30.239999   29.408000    77.408001
192      2048.0        128.0   1.0  20.160001   19.231999    43.968000
193      2048.0        128.0   2.0  21.152001   20.288000    44.096000
194      2048.0        128.0   4.0  23.424000   22.399999    48.384000
195      2048.0        128.0   8.0  28.320000   28.000001    60.479999
196      2048.0        256.0   1.0  20.160001   20.160001    50.112002
197      2048.0        256.0   2.0  21.824000   21.792000    55.103999
198      2048.0        256.0   4.0  24.095999   24.064001    60.768001
199      2048.0        256.0   8.0  29.727999   29.792000    68.928003
200      4096.0          8.0   1.0  20.992000   19.376000   114.576001
201      4096.0          8.0   2.0  23.072001   21.568000   203.615993
202      4096.0          8.0   4.0  27.807999   26.880000   377.952009
203      4096.0          8.0   8.0  36.543999   36.640000   728.096008
204      4096.0         32.0   1.0  21.024000   19.424001    49.024001
205      4096.0         32.0   2.0  23.903999   22.496000    73.632002
206      4096.0         32.0   4.0  28.736001   28.063999   118.303999
207      4096.0         32.0   8.0  38.880002   38.943999   209.503993
208      4096.0         64.0   1.0  21.632001   20.352000    43.327998
209      4096.0         64.0   2.0  24.160000   23.135999    53.280000
210      4096.0         64.0   4.0  30.208001   29.408000    77.264000
211      4096.0         64.0   8.0  40.959999   41.248001   122.560002
212      4096.0        128.0   1.0  21.215999   20.352000    45.952000
213      4096.0        128.0   2.0  23.360001   22.624001    47.775999
214      4096.0        128.0   4.0  28.352000   28.096000    60.288001
215      4096.0        128.0   8.0  37.967999   38.511999    84.352002
216      4096.0        256.0   1.0  21.824000   21.728000    54.912001
217      4096.0        256.0   2.0  24.064001   24.160000    60.479999
218      4096.0        256.0   4.0  29.632000   29.696001    68.928003
219      4096.0        256.0   8.0  35.615999   36.543999    82.208000
220      8192.0          8.0   1.0  23.040000   21.695999   203.968003
221      8192.0          8.0   2.0  27.744001   26.912000   378.176004
222      8192.0          8.0   4.0  36.575999   36.672000   730.368018
223      8192.0          8.0   8.0  54.848000   57.055999  1443.904042
224      8192.0         32.0   1.0  23.808001   22.655999    73.536001
225      8192.0         32.0   2.0  28.704001   27.872000   118.368000
226      8192.0         32.0   4.0  38.784001   38.896000   209.503993
227      8192.0         32.0   8.0  58.432002   60.640000   381.760001
228      8192.0         64.0   1.0  24.160000   23.040000    53.119998
229      8192.0         64.0   2.0  30.176001   29.344000    77.472001
230      8192.0         64.0   4.0  40.832002   41.184001   122.368000
231      8192.0         64.0   8.0  63.072003   65.504000   215.136006
232      8192.0        128.0   1.0  23.391999   22.368001    47.839999
233      8192.0        128.0   2.0  28.320000   28.320000    60.479999
234      8192.0        128.0   4.0  37.951998   38.431998    84.448002
235      8192.0        128.0   8.0  56.384001   59.071999   131.743997
236      8192.0        256.0   1.0  24.095999   24.127999    60.927998
237      8192.0        256.0   2.0  29.664000   29.600000    68.960004
238      8192.0        256.0   4.0  35.583999   36.607999    82.176000
239      8192.0        256.0   8.0  52.480001   55.424001   106.271997
```

This PR:
```
$python ./sgl-kernel/benchmark/bench_moe_align_block_size.py
INFO 07-09 13:51:19 [__init__.py:244] Automatically detected platform cuda.
✅ VLLM implementation works with 64 experts!
✅ SGL and Triton implementations match
✅ SGL and VLLM implementations match

📊 Running performance benchmark for 64 experts...
moe-align-block-size-performance:
     num_tokens  num_experts  topk        SGL  SGL Fusion       Triton
0           1.0          8.0   1.0  16.832000   14.912000    42.431999
1           1.0          8.0   2.0  16.672000   15.072000    42.240001
2           1.0          8.0   4.0  16.864000   15.072000    44.383999
3           1.0          8.0   8.0  16.511999   15.104000    43.327998
4           1.0         32.0   1.0  19.231999   17.503999    42.463999
5           1.0         32.0   2.0  19.264000   17.535999    41.760001
6           1.0         32.0   4.0  19.200001   17.535999    43.200001
7           1.0         32.0   8.0  19.264000   17.568000    44.128001
8           1.0         64.0   1.0  22.655999   20.927999    41.680001
9           1.0         64.0   2.0  22.655999   20.912000    43.200001
10          1.0         64.0   4.0  22.655999   20.959999    43.584000
11          1.0         64.0   8.0  22.592001   20.927999    43.264002
12          1.0        128.0   1.0  20.447999   18.751999    43.584000
13          1.0        128.0   2.0  20.416001   18.784000    42.112000
14          1.0        128.0   4.0  20.447999   18.784000    44.032000
15          1.0        128.0   8.0  20.479999   18.608000    43.232001
…