@yuan-luo 2025-07-10
> Could you also add end2end accuracy test for dsv3 or qwen3 models?

QWEN3:
```
$python3 benchmark/gsm8k/bench_sglang.py --num-questions 200 --parallel 128 --num-shots 8 --port 30000
100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 200/200 [00:11<00:00, 17.07it/s]
Accuracy: 0.725
Invalid: 0.000
Latency: 12.372 s
Output throughput: 4044.128 token/s
```